refact: crud use cases

+++ src/main/java/overengineer/projecthttp/infra/use_case/crud/person/DeletePerson.java
@@ -1,7 +1,8 @@
 package overengineer.projecthttp.infra.use_case.crud.person;
 
 import org.springframework.stereotype.Service;
-import overengineer.projecthttp.domain.PersonGateway;
+import overengineer.projecthttp.domain.person.PersonGateway;
+import overengineer.projecthttp.infra.exception.ApplicationException;
 
 import java.util.UUID;
 
@@ -9,6 +10,9 @@ import java.util.UUID;
 public record DeletePerson(PersonGateway repo) {
 
     public void byId(UUID id) {
+        if (id == null) {
+            throw new ApplicationException("Id cannot be null");
+        }
         repo.deleteById(id);
     }
 }
